pyproject.toml
[project]
name = "dinkum-bio"
version = "0.6.0-dev"
description = "directed interaction network kit for modeling, focused on GRNs"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
"Programming Language :: Python :: Implementation :: CPython",
"License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"
    ]
dependencies = ["ipycanvas==0.13.1", "Pillow==10.1.0", "pandas>=2,<3",
    "nbval==0.11.0", "matplotlib", "lmfit"]

authors = [
  { name="C. Titus Brown" },
]

[metadata]
license = { text = "GNU Affero General Public License v3" }

[tool.pytest.ini_options]
addopts = "--doctest-glob='doc/*.md'"

testpaths = [
  "src/tests",
]
